+++ frontend/src/modules/report/pages/ReportPage.vue
@@ -1,7 +1,7 @@
 <script setup>
-import { ref, onMounted, onBeforeUnmount, computed } from 'vue'
-import Chart from 'chart.js/auto'
+import { ref, onMounted, computed } from 'vue'
 import { getReport } from '../services/report.service'
+import Charts from '../components/Charts.vue'
 
 const data = ref(null)
 
@@ -21,79 +21,9 @@ const bottom10 = computed(() => {
 
 const currency = n => `₩${Number(n || 0).toLocaleString()}`
 
-const barRef = ref(null)
-const lineRef = ref(null)
-let barChart, lineChart
-
-const drawCharts = () => {
-  if (!data.value) return
-  const names = data.value.employees.map(e => e.name)
-  const base = data.value.employees.map(e => e.baseSalary)
-  const raised = data.value.employees.map(e => e.raisedSalary || e.baseSalary)
-  const scores = data.value.employees.map(e => e.score || 0)
-
-  // Bar chart: 이전 연봉 vs 상승 후 연봉
-  if (barChart) barChart.destroy()
-  barChart = new Chart(barRef.value.getContext('2d'), {
-    type: 'bar',
-    data: {
-      labels: names,
-      datasets: [
-        { label: '이전 연봉', data: base, backgroundColor: '#74b9ff' },
-        { label: '상승 후 연봉', data: raised, backgroundColor: '#a29bfe' },
-      ],
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      plugins: {
-        legend: { position: 'top' },
-        tooltip: {
-          callbacks: {
-            label: ctx => `${ctx.dataset.label}: ${currency(ctx.parsed.y)}`,
-          },
-        },
-      },
-      scales: {
-        y: {
-          ticks: { callback: v => currency(v) },
-          beginAtZero: true,
-        },
-      },
-    },
-  })
-
-  // Line chart: 직원별 점수 추이
-  if (lineChart) lineChart.destroy()
-  lineChart = new Chart(lineRef.value.getContext('2d'), {
-    type: 'line',
-    data: {
-      labels: names,
-      datasets: [
-        {
-          label: '직원별 점수 추이',
-          data: scores,
-          fill: true,
-          tension: 0.2,
-          cubicInterpolationMode: 'monotone',
-          borderColor: '#ff7675',
-          backgroundColor: 'rgba(255,118,117,0.2)',
-          pointBackgroundColor: '#ff7675',
-        },
-      ],
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      scales: { y: { suggestedMin: 0, suggestedMax: 100 } },
-    },
-  })
-}
-
 const load = async () => {
   const res = await getReport()
   data.value = res.data.data
-  requestAnimationFrame(() => drawCharts())
 }
 
 onMounted(load)
@@ -135,18 +65,10 @@ onBeforeUnmount(() => {
       </div>
     </div>
 
-    <div class="row g-3">
-      <div class="col-md-6">
-        <div style="height: 320px" class="card p-3">
-          <canvas ref="barRef"></canvas>
-        </div>
-      </div>
-      <div class="col-md-6">
-        <div style="height: 320px" class="card p-3">
-          <canvas ref="lineRef"></canvas>
-        </div>
-      </div>
-    </div>
+    <Charts :names="data.employees.map(e=>e.name)"
+            :base="data.employees.map(e=>e.baseSalary)"
+            :raised="data.employees.map(e=>e.raisedSalary || e.baseSalary)"
+            :scores="data.employees.map(e=>e.score || 0)" />
 
     <div class="row g-3 mt-3">
       <div class="col-md-6">
